Move dm settings folders around
