+++ client/src/game/ui/settings/permissions.vue
